moved scan into separate file
