+++ lib/model/rofolder.go
@@ -8,7 +8,6 @@ package model
 
 import (
 	"fmt"
-	"math/rand"
 	"time"
 
 	"github.com/syncthing/syncthing/lib/sync"
@@ -21,43 +20,6 @@ type roFolder struct {
 	folderId string
 	model    *Model
 	stop     chan struct{}
-}
-
-type rescanRequest struct {
-	subdirs []string
-	err     chan error
-}
-
-// bundle all folder scan activity
-type scan struct {
-	scanInterval time.Duration
-	scanTimer    *time.Timer
-	scanNow      chan rescanRequest
-	scanDelay    chan time.Duration
-}
-
-func (s *scan) rescheduleScan() {
-	if s.scanInterval == 0 {
-		return
-	}
-	// Sleep a random time between 3/4 and 5/4 of the configured interval.
-	sleepNanos := (s.scanInterval.Nanoseconds()*3 + rand.Int63n(2*s.scanInterval.Nanoseconds())) / 4
-	interval := time.Duration(sleepNanos) * time.Nanosecond
-	l.Debugln(s, "next rescan in", interval)
-	s.scanTimer.Reset(interval)
-}
-
-func (s *scan) Scan(subdirs []string) error {
-	req := rescanRequest{
-		subdirs: subdirs,
-		err:     make(chan error),
-	}
-	s.scanNow <- req
-	return <-req.err
-}
-
-func (s *scan) DelayScan(next time.Duration) {
-	s.scanDelay <- next
 }
 
 func newROFolder(model *Model, folderId string, scanInterval time.Duration) *roFolder {
